+++ src/cli/index.ts
@@ -1,16 +1,15 @@
 #!/usr/bin/env node
 
 import { Command } from 'commander';
-import { testsCommand } from './commands/test';
-
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const pkg = require('../../package.json');
+import { testCommand } from './commands/test';
+import { envCommand, getOwnVersion } from './commands/env';
 
 const program = new Command();
 
 program
   .name('bddgen')
-  .description(`Playwright-bdd CLI v${pkg.version}`)
-  .addCommand(testsCommand, { isDefault: true })
+  .description(`Playwright-bdd CLI v${getOwnVersion()}`)
+  .addCommand(testCommand, { isDefault: true })
+  .addCommand(envCommand)
   .addHelpCommand(false)
   .parse();