+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## dev
 * Add [`--tags`](https://github.com/vitalets/playwright-bdd#bddgen-test-or-just-bddgen) option to filter scenarios by tags expression [#36](https://github.com/vitalets/playwright-bdd/issues/36).
+* Add `bddgen env` command to display environment info.
 
 ## 5.0.1
 There are several breaking changes in this release.